Added label RBAC Tests

+++ tests/lib/common/api/pages/base.py
@@ -1,9 +1,9 @@
 import logging
 import inspect
 import httplib
-
-from common.api.schema import validate
 from common.api.pages import Page
+from common.api.pages.labels import Label_Page
+from common.api.schema import validate
 import common.exceptions
 
 
@@ -156,8 +156,12 @@ class Base(Page):
 
     def silent_delete(self):
         '''
-        Delete the object.  If it's already deleted, ignore the error
+        Delete the object. If it's already deleted, ignore the error
         '''
+        # Label pages do not support delete
+        if isinstance(self, Label_Page):
+            return
+
         r = self.api.delete(self.base_url.format(**self.json))
         try:
             return self.handle_request(r)
